Prototype1_July5
This is my attempt at a very bare bones prototype. There is a player and four copies of a basic, exploding enemy. I ran out of time before the player got any attacks. I am not precious about this at all and want it to start getting ideas going. I tried to label everything so it will be easy to throw away later.