I added PointLight2D to LavaAnts and LavaAOE
I added Pointlight2D nodes to the lava ant to try to make it look like their red bulbs were glowing. It might be better to control this with animations, as it looks a little off when the ant looks upward.  I also added a PointLight2D to the LavaAOEs because they looked less bright than the lava ants and that seemed off.