EndScreen
First I zoomed the camera out a little, and then I created an End menu to show how many crystals the player collected, how many enemies died, and how many levels the player made it to. The level counting is not connected anywhere yet since I don't know where the level changing  script it. Also, this scene is not yet connected. The button to go back to title screen is connected though.